Pulling base kpt back in (#4)
* Rename instruction U_504 to HideTextBox

-Rename U_504 class to HideTextBox

-Replace DataBox with proper argument information

* Sort OpCode lists numerically

* Rename U_402 to WAIT

* Rename U_404 to ROTATE_CAMERA

* Rename U_170 to CHANGE_MUSIC

* Update documentation for U_111

* Change U_112 to IntraFileJump

-Rename U_112 to IntraFileJump

-Implement an IInstruction for IntraFileJump

* Change U_113 to InterFileJump

-Rename U_113 to InterFileJump

-Implement an IInstruction for InterFileJump

* Rename U_900 to PLAY_INTERRUPT

* Make Opcode names all caps

-Change InterFileJump and IntraFileJump to match convention set
by pervious instruction names

* Fix InterFileJump bugs

-Previous commmit adding InterFileJump did not update map correctly

-Previous commmit adding InterFileJump did not have correct csproj update

* Implement StCpNumber

-Implement StCpNumber class

-Transition files which previously handled StCp numbers manually to use it

-Expose public constant with header size in St_Header

* Basic outline of jump table management

* Change headers

-Rename old St_Header to StCp_Header

-Introduce new St_Header

The previous St_Header was misleadingly named since it works for
StCp files only. So it has been renamed to StCp_Header. This also
frees up the name St_Header to be repurposed as a generic header
for other files in the St family.

* Create an ElementHelper

-Create element helper with extension function to
get element size

* Move footer to its own IElement

-Create St_Footer class to handle footer

-Create IFooter interface to match existing layout

* Update KPT.csproj

-Previous commit updating headers did not update project file

* Move MakeInstructionBox to ElementHelper

* Add ability to read/write jump tables

-Give JumpTableInterface ability to read/write jump tables

-Correct error in size of JumpTableEntry

* More detailed variable naming for JumpTableEntry

-Convert first DataBox to named variables

* Adjust offset when reading/writing offset table

Turn offsets from header to offsets from start of file
to make working with their location in file streams
easier

* Remove debug string

* StCp update

-Have StCpNumber return its StCp number as a string

-Make StCpNumber of StCp_Header accessible through exposed function

* Add basic implementationf of JumpLabelManager

-Have FileParser check instructions for jumps

-Create JumpLabelManager to mark jump instructions

-Have JumpTableEntry use Jump IDs to help keep track
of jumps

-Have JumpTableInterface expose list of jump table entries

-Create a new IInstruction that wraps around instructions
which are the target of jumps

-Load and save jump table while loading strings

* Change way StCp number creates string of itself

Adding St and Cp to the resulting string should make it
more clear what is actually going on.

* Initialize JumpLabelManger in DumpStrings

FileParser requires JumpLabelManager to be initalized and
DumpStrings needs to use FileParser

* Convert JumpTargetInstructionWrapper to VirtualLabel

Wrapping instructions that were targets of jumps with another
IInstruction was obscuring the interfaces of the wrapped
instruction which was breaking code that relied on being able
to access them. It seemed simpler to just convert
JumpTargetInstructionWrapper to its own standalone IInstruction
than try to expose the interfaces of the wrapped instruction.

* Fix bug with VirtualLabel

VirtualLabel should be inserted before instruction is read not after

For some reason the previous code was capable of passing tests... sometimes

* Make JumpLabelManager non-static

-Make JumpLabelManager and its members/methods non-static

-Have FileParser take JumpLabelManager as an argument

-Make JumpLabelTracking in FileParser optional

Making JumpLabelManager static was in retrospect the
least better option. Originally I did not want to
have to pass it around constantly but having
another static class with arcane initalization
rules and state floating around does not really
seem to be worth it.

* Fix outdated variable name

Variable wrappedInstruction should have been changed
to virtualLabel during the conversion of
JumpTargetInstructionWrapper to VirtualLabel

* Change StCp number ToString to use hex

As it turns out, the file names directly correspond to
the numbers in the headers printed as hexadecimal

Co-authored-by: dl471 <annamika@bk.ru>
Co-authored-by: dl471 <32369415+dl471@users.noreply.github.com>